[Feature] Extend Pod Security context
