+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Improvement) Unify K8S Error Handling
 - (Feature) Remove stuck Pods
 - (Bugfix) Fix Go routine leak
+- (Feature) Extend Pod Security context
 
 ## [1.2.19](https://github.com/arangodb/kube-arangodb/tree/1.2.19) (2022-10-05)
 - (Bugfix) Prevent changes when UID is wrong
