+++ pkg/apis/deployment/v2alpha1/server_group_spec.go
@@ -157,102 +157,6 @@ type ServerGroupSpec struct {
 	IndexMethod *ServerGroupIndexMethod `json:"indexMethod,omitempty"`
 }
 
-// ServerGroupSpecSecurityContext contains specification for pod security context
-type ServerGroupSpecSecurityContext struct {
-	// DropAllCapabilities specifies if capabilities should be dropped for this pod containers
-	//
-	// Deprecated: This field is added for backward compatibility. Will be removed in 1.1.0.
-	DropAllCapabilities *bool `json:"dropAllCapabilities,omitempty"`
-	// AddCapabilities add new capabilities to containers
-	AddCapabilities []core.Capability `json:"addCapabilities,omitempty"`
-
-	AllowPrivilegeEscalation *bool  `json:"allowPrivilegeEscalation,omitempty"`
-	Privileged               *bool  `json:"privileged,omitempty"`
-	ReadOnlyRootFilesystem   *bool  `json:"readOnlyRootFilesystem,omitempty"`
-	RunAsNonRoot             *bool  `json:"runAsNonRoot,omitempty"`
-	RunAsUser                *int64 `json:"runAsUser,omitempty"`
-	RunAsGroup               *int64 `json:"runAsGroup,omitempty"`
-
-	SupplementalGroups []int64 `json:"supplementalGroups,omitempty"`
-	FSGroup            *int64  `json:"fsGroup,omitempty"`
-}
-
-// GetDropAllCapabilities returns flag if capabilities should be dropped
-//
-// Deprecated: This function is added for backward compatibility. Will be removed in 1.1.0.
-func (s *ServerGroupSpecSecurityContext) GetDropAllCapabilities() bool {
-	if s == nil {
-		return true
-	}
-
-	if s.DropAllCapabilities == nil {
-		return true
-	}
-
-	return *s.DropAllCapabilities
-}
-
-// GetAddCapabilities add capabilities to pod context
-func (s *ServerGroupSpecSecurityContext) GetAddCapabilities() []core.Capability {
-	if s == nil {
-		return nil
-	}
-
-	if s.AddCapabilities == nil {
-		return nil
-	}
-
-	return s.AddCapabilities
-}
-
-// NewSecurityContext creates new pod security context
-func (s *ServerGroupSpecSecurityContext) NewPodSecurityContext() *core.PodSecurityContext {
-	if s == nil {
-		return nil
-	}
-
-	if s.FSGroup == nil && len(s.SupplementalGroups) == 0 {
-		return nil
-	}
-
-	return &core.PodSecurityContext{
-		SupplementalGroups: s.SupplementalGroups,
-		FSGroup:            s.FSGroup,
-	}
-}
-
-// NewSecurityContext creates new security context
-func (s *ServerGroupSpecSecurityContext) NewSecurityContext() *core.SecurityContext {
-	r := &core.SecurityContext{}
-
-	if s != nil {
-		r.AllowPrivilegeEscalation = s.AllowPrivilegeEscalation
-		r.Privileged = s.Privileged
-		r.ReadOnlyRootFilesystem = s.ReadOnlyRootFilesystem
-		r.RunAsNonRoot = s.RunAsNonRoot
-		r.RunAsUser = s.RunAsUser
-		r.RunAsGroup = s.RunAsGroup
-	}
-
-	capabilities := &core.Capabilities{}
-
-	if s.GetDropAllCapabilities() {
-		capabilities.Drop = []core.Capability{
-			"ALL",
-		}
-	}
-
-	if caps := s.GetAddCapabilities(); caps != nil {
-		capabilities.Add = []core.Capability{}
-
-		capabilities.Add = append(capabilities.Add, caps...)
-	}
-
-	r.Capabilities = capabilities
-
-	return r
-}
-
 // ServerGroupProbesSpec contains specification for probes for pods of the server group
 type ServerGroupProbesSpec struct {
 	// LivenessProbeDisabled if true livenessProbes are disabled
